# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aryventory Sales Management API",
        "description": "API for managing sales executives, leads, meetings, shifts, and business operations",
        "contact": {
            "email": "support@aryventory.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        },
        {
            "url": "https://api.aryventory.com/api",
            "description": "Production server"
        }
    ],
    "paths": {
        "/api/dashboard/overview": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard overview",
                "description": "Retrieve comprehensive dashboard overview with shift counts, revenue metrics, and performance data",
                "operationId": "getDashboardOverview",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for data filtering (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for data filtering (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "region",
                        "in": "query",
                        "description": "Filter by region/group ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "manager_id",
                        "in": "query",
                        "description": "Filter by manager ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dashboard overview retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "shift_overview": {
                                            "properties": {
                                                "total_shifts_started": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "active_shifts": {
                                                    "type": "integer",
                                                    "example": 8
                                                },
                                                "completed_shifts": {
                                                    "type": "integer",
                                                    "example": 17
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "revenue_metrics": {
                                            "properties": {
                                                "by_executive": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "user_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "user_name": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            },
                                                            "designation": {
                                                                "type": "string",
                                                                "example": "Sales Executive"
                                                            },
                                                            "group": {
                                                                "type": "string",
                                                                "example": "North Region"
                                                            },
                                                            "completed_leads": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "estimated_revenue": {
                                                                "type": "integer",
                                                                "example": 250000
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "by_manager": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "manager_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "manager_name": {
                                                                "type": "string",
                                                                "example": "Jane Smith"
                                                            },
                                                            "completed_leads": {
                                                                "type": "integer",
                                                                "example": 15
                                                            },
                                                            "estimated_revenue": {
                                                                "type": "integer",
                                                                "example": 750000
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "by_region": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "region_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "region_name": {
                                                                "type": "string",
                                                                "example": "North Region"
                                                            },
                                                            "completed_leads": {
                                                                "type": "integer",
                                                                "example": 20
                                                            },
                                                            "estimated_revenue": {
                                                                "type": "integer",
                                                                "example": 1000000
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total_completed_leads": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "total_estimated_revenue": {
                                                    "type": "integer",
                                                    "example": 1250000
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "performance_metrics": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "designation": {
                                                        "type": "string",
                                                        "example": "Sales Executive"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "Sales Executive"
                                                    },
                                                    "group": {
                                                        "type": "string",
                                                        "example": "North Region"
                                                    },
                                                    "leads_created": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "leads_completed": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "meetings_conducted": {
                                                        "type": "integer",
                                                        "example": 8
                                                    },
                                                    "shifts_worked": {
                                                        "type": "integer",
                                                        "example": 20
                                                    },
                                                    "conversion_rate": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 50
                                                    },
                                                    "target": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "target_achievement": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 33.33
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "date_range": {
                                            "properties": {
                                                "start_date": {
                                                    "type": "string",
                                                    "example": "2024-01-01"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "example": "2024-01-31"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/leads-bulk": {
            "delete": {
                "tags": [
                    "Leads"
                ],
                "summary": "Bulk delete leads",
                "description": "Delete multiple leads at once",
                "operationId": "bulkDeleteLeads",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lead_ids"
                                ],
                                "properties": {
                                    "lead_ids": {
                                        "description": "Array of lead IDs to delete",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Leads deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully deleted 3 lead(s)"
                                        },
                                        "deleted_count": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "lead_ids": [
                                                    "The lead ids field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/sales-executive/lead-counts-by-status": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Get lead counts by status",
                "description": "Get lead counts grouped by status for the logged-in sales executive",
                "operationId": "getLeadCountsByStatus",
                "responses": {
                    "200": {
                        "description": "Lead counts by status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "lead_status": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "status_name": {
                                                "type": "string",
                                                "example": "New"
                                            },
                                            "total": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": {
                "sanctum": []
            }
        }
    },
...
}